+++ publish-common/src/main/java/com/ericsson/eiffel/remrem/publish/helper/RabbitMqProperties.java
@@ -540,7 +540,20 @@ public class RabbitMqProperties {
         if ((rabbitConnection == null || !rabbitConnection.isOpen())) {
             createRabbitMqConnection();
         }
-        return rabbitChannels.get(random.nextInt(rabbitChannels.size()));
+        for (Channel channel : rabbitChannels) {
+            if (channel.isOpen()) {
+                return channel;
+            }
+        }
+        try {
+            Channel channel = rabbitConnection.createChannel();
+            channel.confirmSelect();
+            rabbitChannels.add(channel);
+            return channel;
+        } catch (IOException e) {
+            log.error(e.getMessage(), e);
+            throw new RemRemPublishException("Failed to create new channel for Rabbitmq :: ",
+                    factory, e);
+        }
     }
-
 }